pyproject.toml
[project]
name = "openscientist"
version = "0.1.0"
description = "OpenScientist - Scientific Hypothesis Agent for Novel Discovery"
authors = [
    {name = "Justin Reese", email = "justinreese@lbl.gov"}
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    # Web framework
    "nicegui>=3.7.1",

    # Core scientific computing
    "pandas>=3.0.0",
    "numpy>=2.2.0,<2.4",
    "scipy>=1.15.0",
    "scikit-learn>=1.6.0",

    # File format support
    "python-magic>=0.4.27",  # MIME type detection
    "openpyxl>=3.1.0",       # Excel .xlsx support
    "pyarrow>=14.0.0",       # Parquet support
    "pymupdf>=1.24.0",       # PDF text extraction
    "python-docx>=1.1.0",    # Word document text extraction

    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",

    # Statistics
    "statsmodels>=0.14.0",

    # Domain-specific
    "networkx>=3.0",

    # Single-cell genomics
    "scanpy>=1.12.0",
    "anndata>=0.12.0",
    "h5py>=3.10.0",

    # API clients
    "anthropic>=0.78.0",
    "requests>=2.31.0",

    # MCP server
    "mcp[cli]>=1.0.0",

    # Environment & config
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",

    # Authentication (OAuth)
    "authlib>=1.3.0",
    "httpx>=0.27.0",  # Required by authlib for async OAuth
    "python-jose[cryptography]>=3.3.0",  # JWT handling
    "bcrypt>=4.0.0",  # Password hashing

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",  # PostgreSQL async driver
    "alembic>=1.13.0",  # Database migrations

    # REST API
    "slowapi>=0.1.9",  # Rate limiting for FastAPI

    # Type hints
    "typing-extensions>=4.8.0",

    # Cloud providers (for cost tracking and authentication)
    "google-cloud-bigquery>=3.11.0",  # Vertex AI
    "boto3>=1.34.0",  # AWS Bedrock
    "azure-identity>=1.15.0",  # Azure Foundry authentication
    "azure-mgmt-costmanagement>=4.0.0",  # Azure Foundry cost tracking

    # PDF generation (lightweight, pure Python)
    "fpdf2>=2.7.0",
    "markdown>=3.5.0",

    # Berkeley Lab Data Lakehouse
    "linkml-store[dremio]>=0.2.0",

    # Container isolation
    "docker>=7.0.0",

    # Claude Agent SDK (required – powers the agent execution path)
    "claude-agent-sdk>=0.1.37",

    # SPARQL queries against remote endpoints
    "SPARQLWrapper>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.lint.per-file-ignores]
"src/openscientist/database/migrations/versions/*" = ["E402"]  # Alembic auto-generated imports after docstrings

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: marks tests as integration tests (require Docker and executor image)",
]

[tool.coverage.run]
source = ["src/openscientist"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
    "*/site-packages/*",
]
relative_files = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@overload",
    "except ImportError:",
]
# Minimum coverage threshold
fail_under = 60

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

[tool.coverage.paths]
source = ["src/openscientist"]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-requests>=2.31.0",
    "pandas-stubs>=3.0.0",
    "types-seaborn>=0.13.0",
    "types-openpyxl>=3.1.0",
    "types-fpdf2>=2.7.0",
    "types-docker>=7.0.0",
    "pre-commit>=3.5.0",
    "testcontainers[postgres]>=4.0.0",
    "cairosvg>=2.8.2",
]
